# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobShield AI - Cybersecurity scam detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Type of content being analyzed
 */
export type AnalysisInputContentType = typeof AnalysisInputContentType[keyof typeof AnalysisInputContentType];


export const AnalysisInputContentType = {
  recruiter_message: 'recruiter_message',
  job_posting: 'job_posting',
  linkedin_message: 'linkedin_message',
  email: 'email',
  other: 'other',
} as const;

export interface AnalysisInput {
  /**
     * The text content to analyze (recruiter message, job posting, etc.)
     * @minLength 10
     */
  content: string;
  /** Type of content being analyzed */
  contentType: AnalysisInputContentType;
  /**
     * Optional source label (e.g. company name, platform)
     * @nullable
     */
  source?: string | null;
}

export type RedFlagSeverity = typeof RedFlagSeverity[keyof typeof RedFlagSeverity];


export const RedFlagSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface RedFlag {
  id: string;
  category: string;
  description: string;
  severity: RedFlagSeverity;
  matched: boolean;
  /** @nullable */
  matchedText?: string | null;
}

export type AnalysisContentType = typeof AnalysisContentType[keyof typeof AnalysisContentType];


export const AnalysisContentType = {
  recruiter_message: 'recruiter_message',
  job_posting: 'job_posting',
  linkedin_message: 'linkedin_message',
  email: 'email',
  other: 'other',
} as const;

/**
 * low: 0-25, medium: 26-50, high: 51-75, critical: 76-100
 */
export type AnalysisThreatLevel = typeof AnalysisThreatLevel[keyof typeof AnalysisThreatLevel];


export const AnalysisThreatLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Analysis {
  id: number;
  content: string;
  contentType: AnalysisContentType;
  /** @nullable */
  source?: string | null;
  /**
     * Risk score from 0 (safe) to 100 (definitely scam)
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  /** low: 0-25, medium: 26-50, high: 51-75, critical: 76-100 */
  threatLevel: AnalysisThreatLevel;
  redFlags: RedFlag[];
  /** Actionable recommendations based on findings */
  recommendations: string[];
  /** Human-readable explanation of the findings */
  summary: string;
  createdAt: string;
}

export type ThreatLevelCountLevel = typeof ThreatLevelCountLevel[keyof typeof ThreatLevelCountLevel];


export const ThreatLevelCountLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface ThreatLevelCount {
  level: ThreatLevelCountLevel;
  count: number;
}

export interface TopRedFlag {
  category: string;
  count: number;
}

export interface Stats {
  totalScans: number;
  avgRiskScore: number;
  /** Analyses with low threat level */
  safeCount: number;
  /** Analyses with high or critical threat level */
  scamCount: number;
  threatLevelCounts: ThreatLevelCount[];
  topRedFlags: TopRedFlag[];
}

export type IndicatorSeverity = typeof IndicatorSeverity[keyof typeof IndicatorSeverity];


export const IndicatorSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Indicator {
  id: string;
  category: string;
  name: string;
  description: string;
  severity: IndicatorSeverity;
  examples: string[];
}

export type ListAnalysesParams = {
limit?: number;
offset?: number;
threatLevel?: ListAnalysesThreatLevel;
};

export type ListAnalysesThreatLevel = typeof ListAnalysesThreatLevel[keyof typeof ListAnalysesThreatLevel];


export const ListAnalysesThreatLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;