# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobShield AI - Cybersecurity scam detection API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisContentType } from './analysisContentType';
import type { AnalysisThreatLevel } from './analysisThreatLevel';
import type { RedFlag } from './redFlag';

export interface Analysis {
  id: number;
  content: string;
  contentType: AnalysisContentType;
  /** @nullable */
  source?: string | null;
  /**
     * Risk score from 0 (safe) to 100 (definitely scam)
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  /** low: 0-25, medium: 26-50, high: 51-75, critical: 76-100 */
  threatLevel: AnalysisThreatLevel;
  redFlags: RedFlag[];
  /** Actionable recommendations based on findings */
  recommendations: string[];
  /** Human-readable explanation of the findings */
  summary: string;
  createdAt: Date;
}