# lib/api-zod/src/generated/types/indicator.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobShield AI - Cybersecurity scam detection API
 * OpenAPI spec version: 0.1.0
 */
import type { IndicatorSeverity } from './indicatorSeverity';

export interface Indicator {
  id: string;
  category: string;
  name: string;
  description: string;
  severity: IndicatorSeverity;
  examples: string[];
}