# lib/api-zod/src/generated/types/redFlag.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobShield AI - Cybersecurity scam detection API
 * OpenAPI spec version: 0.1.0
 */
import type { RedFlagSeverity } from './redFlagSeverity';

export interface RedFlag {
  id: string;
  category: string;
  description: string;
  severity: RedFlagSeverity;
  matched: boolean;
  /** @nullable */
  matchedText?: string | null;
}