# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobShield AI - Cybersecurity scam detection API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatLevelCount } from './threatLevelCount';
import type { TopRedFlag } from './topRedFlag';

export interface Stats {
  totalScans: number;
  avgRiskScore: number;
  /** Analyses with low threat level */
  safeCount: number;
  /** Analyses with high or critical threat level */
  scamCount: number;
  threatLevelCounts: ThreatLevelCount[];
  topRedFlags: TopRedFlag[];
}