# lib/api-zod/src/generated/types/threatLevelCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobShield AI - Cybersecurity scam detection API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatLevelCountLevel } from './threatLevelCountLevel';

export interface ThreatLevelCount {
  level: ThreatLevelCountLevel;
  count: number;
}